+++ src/main/java/com/github/jnhyperion/hyperrobotframeworkplugin/psi/dto/KeywordDto.java
@@ -3,11 +3,16 @@ package com.github.jnhyperion.hyperrobotframeworkplugin.psi.dto;
 import com.github.jnhyperion.hyperrobotframeworkplugin.psi.element.DefinedKeyword;
 import com.github.jnhyperion.hyperrobotframeworkplugin.psi.element.DefinedParameter;
 import com.github.jnhyperion.hyperrobotframeworkplugin.psi.element.DefinedVariable;
+import com.github.jnhyperion.hyperrobotframeworkplugin.psi.ref.PyElementDeprecatedVisitor;
+import com.github.jnhyperion.hyperrobotframeworkplugin.psi.ref.PyElementParentTraversalVisitor;
 import com.github.jnhyperion.hyperrobotframeworkplugin.psi.util.PatternBuilder;
 import com.github.jnhyperion.hyperrobotframeworkplugin.psi.util.PatternUtil;
 import com.github.jnhyperion.hyperrobotframeworkplugin.psi.util.ReservedVariable;
 import com.intellij.psi.PsiElement;
 import com.jetbrains.python.psi.PyBoolLiteralExpression;
+import com.jetbrains.python.psi.PyElement;
+import com.jetbrains.python.psi.PyElementVisitor;
+import com.jetbrains.python.psi.PyExpression;
 import com.jetbrains.python.psi.PyNoneLiteralExpression;
 import com.jetbrains.python.psi.PyParameter;
 import com.jetbrains.python.psi.PyReferenceExpression;
@@ -32,11 +37,13 @@ public class KeywordDto implements DefinedKeyword {
     private final boolean args;
     private final Pattern namePattern;
     private final Collection<DefinedParameter> parameters;
+    private final boolean deprecated;
 
     public KeywordDto(@NotNull PsiElement reference, @NotNull String namespace, @NotNull String name) {
         this(reference, namespace, name, null);
     }
 
+    // Used for user defined keywords
     public KeywordDto(@NotNull PsiElement reference, @NotNull String name, @Nullable Collection<DefinedParameter> parameters) {
         this(reference, "", name, parameters);
     }
@@ -46,20 +53,28 @@ public class KeywordDto implements DefinedKeyword {
     }
 
     public KeywordDto(@NotNull PsiElement reference,
-                       @NotNull String namespace,
-                       @NotNull String name,
-                       Collection<DefinedParameter> parameters,
-                       Object... ignored) { // Object... ignored is just a trick to avoid constructor clashing
+                      @NotNull String namespace,
+                      @NotNull String name,
+                      Collection<DefinedParameter> parameters,
+                      Object... ignored) { // Object... ignored is just a trick to avoid constructor clashing
         this.reference = reference;
         this.name = PatternUtil.functionToKeyword(name).trim();
         this.namePattern = Pattern.compile(PatternBuilder.parseNamespace(namespace, PatternUtil.keywordToFunction(this.name)), Pattern.CASE_INSENSITIVE);
         this.args = parameters != null && !parameters.isEmpty();
         this.parameters = parameters;
+        if (reference instanceof PyElement) {
+            PyElementDeprecatedVisitor deprecationVisitor = new PyElementDeprecatedVisitor();
+            PyElementVisitor pyElementParentTraversalVisitor = new PyElementParentTraversalVisitor(deprecationVisitor);
+            reference.accept(pyElementParentTraversalVisitor);
+            deprecated = deprecationVisitor.isDeprecated();
+        } else {
+            deprecated = false;
+        }
     }
 
     @Override
     public final String getKeywordName() {
-        return this.name;
+        return name;
     }
 
     @Override
@@ -97,15 +112,13 @@ public class KeywordDto implements DefinedKeyword {
     @Nullable
     private static String extractDefaultValue(PyParameter parameter) {
         String defaultValue;
-        if (parameter.getDefaultValue() instanceof PyStringLiteralExpression expression) {
-            defaultValue = expression.getStringValue();
-        } else if (parameter.getDefaultValue() instanceof PyBoolLiteralExpression expression) {
-            defaultValue = expression.getValue() ? ReservedVariable.TRUE.getVariable() : ReservedVariable.FALSE.getVariable();
-        } else if (parameter.getDefaultValue() instanceof PyReferenceExpression) {
-            defaultValue = parameter.getDefaultValue().getName();
-        } else if (parameter.getDefaultValue() instanceof PyNoneLiteralExpression) {
-            defaultValue = ReservedVariable.NONE.getVariable();
-        } else {
+        PyDefaultValueExtractor pyDefaultValueExtractor = new PyDefaultValueExtractor();
+        PyExpression defaultValueElement = parameter.getDefaultValue();
+        if (defaultValueElement != null) {
+            defaultValueElement.accept(pyDefaultValueExtractor);
+        }
+        defaultValue = pyDefaultValueExtractor.defaultValue;
+        if (defaultValue == null) {
             defaultValue = parameter.getDefaultValueText();
         }
         return defaultValue;
@@ -134,15 +147,19 @@ public class KeywordDto implements DefinedKeyword {
         for (Object argument : arguments) {
             if (argument instanceof PyParameter parameter) {
                 String name = parameter.getName();
-                //noinspection UnstableApiUsage
                 if (!parameter.isSelf()) {
                     argumentNames.add(name);
                 }
-            } else if (argument instanceof DefinedVariable) {
-                argumentNames.add(((DefinedVariable) argument).getLookup());
+            } else if (argument instanceof DefinedVariable definedVariable) {
+                argumentNames.add(definedVariable.getLookup());
             }
         }
         return " (" + String.join(", ", argumentNames) + ")";
+    }
+
+    @Override
+    public boolean isDeprecated() {
+        return deprecated;
     }
 
     @Override
@@ -164,5 +181,30 @@ public class KeywordDto implements DefinedKeyword {
     @Override
     public String toString() {
         return this.name;
+    }
+
+    private static class PyDefaultValueExtractor extends PyElementVisitor {
+
+        private String defaultValue;
+
+        @Override
+        public void visitPyStringLiteralExpression(@NotNull PyStringLiteralExpression node) {
+            defaultValue = node.getStringValue();
+        }
+
+        @Override
+        public void visitPyBoolLiteralExpression(@NotNull PyBoolLiteralExpression node) {
+            defaultValue = node.getValue() ? ReservedVariable.TRUE.getVariable() : ReservedVariable.FALSE.getVariable();
+        }
+
+        @Override
+        public void visitPyReferenceExpression(@NotNull PyReferenceExpression node) {
+            defaultValue = node.getName();
+        }
+
+        @Override
+        public void visitPyNoneLiteralExpression(@NotNull PyNoneLiteralExpression node) {
+            defaultValue = ReservedVariable.NONE.getVariable();
+        }
     }
 }