+++ src/main/java/com/github/jnhyperion/hyperrobotframeworkplugin/psi/element/DefinedKeyword.java
@@ -17,4 +17,6 @@ public interface DefinedKeyword {
    PsiElement reference();
 
    String getArgumentsDisplayable();
+
+   boolean isDeprecated();
 }